chore: simplify package.json traversal
This replaces `find-package-up` with `empathic`, a faster and smaller library
for generically finding files upwards.

As well as this, it changes the type of the loaded `package.json` to be a
`Record<string, unknown>`. Previously, it had some normalization happen which we
don't actually rely on (stuff npm cares about, since it was using the same
normalization npm uses). This seems OTT for us loading a couple of keys out of
the file.