feat(non-interactive): Add logging for turn number, thought, and tools
For \#4416, this change adds output for thoughts, tool calling, and
results in non-interactive (-y -p) mode.

This will improve the debuggability when using gemini-cli for a batch
process so that the outputs can be collected for later analysis.

The change introduces the following format for the non-interactive mode:

```markdown
\# Turn 1
> Thought: subject
> description
> Thought: subject 2
> description line 1
> description line 2
✦ response text
  mutliple line response
$ Calling tool: glob
toolResponse.resultDisplay

\# Turn 2
...
```

The format is human-readable and kind of similar to interactive mode,
while still able to be processed by a machine (e.g., `awk` can differentiate
the different parts of the output via leading characters "\#", "\>", "✦", and
"$").