feat(shell): Add `preapprovedShellCommandRegexes` to settings.json.
This gives users fine-grained control over preapproved shell commands.  In particular, it avoids problems with the interactive approval, which only keys off of the root command, as well as issues with coreTools, which can only match a prefix and also require an exhaustive list of tools in order to not break basic functionality.