fix(temporary): Don't run ctrl+c exit test on windows
This test is failing on every PR and every run against main. Since we
know this is broken, let's remove the noise and fix it. By allowing it
to continue failing, we are potentially covering up new problems being
introduced.

We should definitely follow up on this asap, either fixing it to pass on
windows, determining that it doesn't need to (does ctrl c on windows
even make sense? i'm not sure, it's been a while XD - does it copy?), or
removing it if it's not necessary. In the meantime let's remove the
noise.